% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanCorpus.R
\name{CleanCorpus}
\alias{CleanCorpus}
\title{Performs a basic corpus cleanse}
\usage{
CleanCorpus(corpus, CapitalAfterPunctuation = FALSE, alltolower = FALSE)
}
\arguments{
\item{corpus}{a tm corpus.}

\item{n}{scalar integer. The length of each chunk to sample.}

\item{k}{scalar integer. The number of samples to take in each text.}

\item{th}{can be 0 or 1. The intertextual distance threshold.}
}
\value{
a list comprehending both the k sampled intertextual distances and the final sampled intertextual distance.
}
\description{
This function calculates the intertextual distance in a corpus.
}
\details{
No details
}
\examples{
library(tm)
corpus <- c("This is a dirty  text (with a double space). 
   There's also a proper name: Lisa and some RANdoM capITal LETTers? 
   Also other capital examples! With other punctuation symbols.",
   "This is another dirty text, with some punctuation.")
corpus <- VCorpus(VectorSource(corpus),readerControl = list(language="ita"))
}
