% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExploreCorpus.R
\name{ExploreCorpus}
\alias{ExploreCorpus}
\title{Prints out corpus-level lexicographic measures}
\usage{
ExploreCorpus(x)
}
\arguments{
\item{x}{a contingency matrix.}
}
\description{
This function saves a data frame with the main corpus-level lexicographic measures.
}
\details{
No details
}
\examples{
library(tm)
corpus <- c("This is a text, with 9 words in total","This is another text, this time with 11 words in total")
corpus <- VCorpus(VectorSource(corpus))
contingencymatrix <- TermDocumentMatrix(corpus)
ExploreCorpus(contingencymatrix)
}
\seealso{
\code{\link{ExploreTexts}}
}
