% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExporeTexts.R
\name{ExploreTexts}
\alias{ExploreTexts}
\title{Prints out text-level lexicographic measures}
\usage{
ExploreTexts(x)
}
\arguments{
\item{x}{a tm corpus.}
}
\description{
This function saves a data frame with the main lexicographic measures about each text in the corpus.
}
\details{
This function reads as a input both a Term-Document Matrix or a Document-Term Matrix.
}
\examples{
library(tm)
corpus <- c("This is a text, with 9 words in total","This is another text, this time with 11 words in total")
corpus <- VCorpus(VectorSource(corpus))
contingencymatrix <- TermDocumentMatrix(corpus)
ExploreTexts(contingencymatrix)
contingencymatrix <- DocumentTermMatrix(corpus)
ExploreTexts(contingencymatrix)
}
\seealso{
\code{\link{ExploreCorpus}}
}
