% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FromTalTacToTxt.R
\name{FromTalTacToTxt}
\alias{FromTalTacToTxt}
\title{TalTac files conversion}
\usage{
FromTalTacToTxt(ttcorpus)
}
\arguments{
\item{ttcorpus}{a character schalar or vector (e.g. you may want to import a TalTac Corpus using readLines)}
}
\value{
a list, comprehensive of a character vector (the texts) and, optionally, the metadata matrix.
}
\description{
Convert TalTac-like corpora to plain txt files
}
\details{
You can use this function to convert a TalTac-like formatted corpus to a number of texts and the relative metadata matrix.
Please note that you can use TalTac to perform this conversion, if you have a TalTac license ;).
Please also note that importing the TalTac corpus with readLines command is the only tested approach, as it grants a new "endline" charachter for each text.
}
\examples{
library(tm)
TalTacCorpus <- file("TalTacCorpus.data","w")
cat("****T1 *V1=a *V2=b", file = TalTacCorpus,  sep = "\n")
cat("This is a text, with 9 words in total.", file = TalTacCorpus, sep = "\n")
cat("****T2 *V1=c *V2=d", file = TalTacCorpus, sep = "\n")
cat(" This is another text, this time with 11 words in total", file = TalTacCorpus,  sep = "\n")
close(TalTacCorpus)
x <- readLines("TalTacCorpus.data")
unlink("TalTacCorpus.data")
txtCorpus <- FromTalTacToTxt(TalTacCorpus)
corpus <- VCorpus(VectorSource(txtCorpus))
}
\seealso{
\code{\link{FromTxtToTalTac}}
}
