% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntDist.R
\name{IntDist}
\alias{IntDist}
\title{Intertextual matrix computation}
\usage{
IntDist(x, th)
}
\arguments{
\item{x}{a word-document contingency matrix.}

\item{th}{can be 0, 1. Sets the threshold for the hapax to be considered (0) or not (1).}
}
\value{
a distance object.
}
\description{
This function calculates the intertextual distance in a corpus.
}
\details{
No details
}
\examples{
library(tm)
texts <- c("This is a text","This is another text")
corpus <- VCorpus(VectorSource(texts))
contmat <- TermDocumentMatrix(corpus)
intdist(contmat,th=0)
}
