% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntDist_WithLoadings.R
\name{IntDistWL}
\alias{IntDistWL}
\title{Intertextual matrix computation}
\usage{
IntDistWL(x, th, ntops = 10)
}
\arguments{
\item{x}{a word-document contingency matrix.}

\item{th}{can be 0, 1. Sets the threshold for the hapax to be considered (0) or not (1).}

\item{ntops}{a scalar integer. The number of most influencing words to be returned.}
}
\value{
a list made "dist" (the intertextual distance), "tops" (most influencing words) and "downs" (less influencing words).
}
\description{
This function calculates the intertextual distance in a corpus.
}
\details{
No details
}
\examples{
library(tm)
texts <- c("This is a text","This is another text")
corpus <- VCorpus(VectorSource(texts))
contmat <- TermDocumentMatrix(corpus)
IntDistWL(contmat,th=0, ntops = 2)
}
