% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeAllTexts.R
\name{MergeAllTexts}
\alias{MergeAllTexts}
\title{Function that merges several texts together}
\usage{
MergeAllTexts(x, collapsevalue = " ")
}
\arguments{
\item{x}{a tm corpus.}
}
\description{
This function reads a corpus as main input, and merges all texts together to obtain a unique text from the corpus.
}
\details{
This function reads a corpus and returns a character string. If needed, the character string should be coherced again to a corpus.
}
\examples{
library(tm)
texts <- c("This is a text","This is another text")
corpus <- VCorpus(VectorSource(texts))
MergedText <- MergeAllTexts(corpus)
MergedCorpus <- VCorpus(VectorSource(MergedText))
MergedCorpus[["1"]]$content
}
