% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeTexts.R
\name{MergeTexts}
\alias{MergeTexts}
\title{Corpus texts merging}
\usage{
MergeTexts(x, group, gStyle, gSize)
}
\arguments{
\item{x}{a tm corpus.}

\item{gStyle}{can be "Factor", "SameNumber", "SameTokens". See Details.}

\item{gSize}{a scalar integer. Is the number of final texts to have in each final group, or the number of tokens for each text depending on gStyle}

\item{gFactor}{a factor or character vector. Contains the category of each text.}
}
\value{
a distance object.
}
\description{
This function can aggregate together texts, with some options available. Useful to analyze a large number of short texts.
}
\details{
#' Used in a corpus where short texts can be grouped based on some characteristics (author, time, ...), this function groups together the texts to form longer, more analyzable texts.
This function can be used with two grouping styles: fixed number of final texts for each group, and equal number of tokens for each final text.
The "SameNumber" option merges texts in order to have the same number of texts for each group. Allocation of texts is optimized in order to have a similar number of tokens in each macro-text.
The "SameTokens" option merges texts in order to have a different number of texts in the same group, but with an optimized balanced number of tokens.
}
\examples{
library(tm)
texts <- c("This is a text for group A","This is another text for group A",
"This is the third text for group A","This is the last text for group A",
"This is a text for group B","This is another text for group B",
"This is the third text for group B","This is the last text for group B")
corpus <- VCorpus(VectorSource(texts))
group <- c("A","A","A","A","B","B","B","B")
mt <- MergeTexts(corpus, group =group, gStyle = "SameNumber", gSize = 2)
mt[[1]]$content
mt[[2]]$content
mt[[3]]$content
mt[[4]]$content
}
