% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampledIntDist.R
\name{SampledIntDist}
\alias{SampledIntDist}
\title{Sampled Intertextual Distance Calculation}
\usage{
SampledIntDist(corpus, n, k, th = 0)
}
\arguments{
\item{corpus}{a tm corpus.}

\item{n}{scalar integer. The length of each chunk to sample.}

\item{k}{scalar integer. The number of samples to take in each text.}

\item{th}{can be 0 or 1. The intertextual distance threshold.}
}
\value{
a list comprehending both the k sampled intertextual distances and the final sampled intertextual distance.
}
\description{
This function calculates the intertextual distance in a corpus.
}
\details{
No details
}
\examples{
library(tm)
texts <- c("This is a text, with 9 words in total","This is another text, this time with 11 words in total")
corpus <- VCorpus(VectorSource(texts))
campdist(corpus,n=6,k=2,th=0)
}
\seealso{
\code{\link{IntDist}}
}
