% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitCorpus.R
\name{SplitCorpus}
\alias{SplitCorpus}
\title{Function that splits each text of corpus in sub-texts}
\usage{
SplitCorpus(x, size, titles)
}
\arguments{
\item{x}{a tm corpus.}

\item{size}{a scalar integer. The size of the final chunks.}

\item{titles}{a character vector. An identifier for corpus texts. If NULL, meta = id is taken.}
}
\description{
This function reads a corpus as main input, splits each text in a fixed size number of sub-texts. 
The number of sub-texts is based on shorter text length. So the size parameter has to be chosen carefully in order to maximize the number of chunks that will be extracted.
}
\details{
This function reads a corpus and returns another corpus.
}
\examples{
library(tm)
texts <- data.frame(doc_id = c("Text A","Text B"), 
text = c("This is a text with 7 words.",
   "This is another text, this time with 8 words."))
corpus <- VCorpus(VectorSource(texts))
SplittedCorpus <- SplitCorpus(corpus, size = 2, titles = c("First text","Second text"))
contMat <- TermDocumentMatrix(SplittedCorpus, control = list(wordLengths=c(0,Inf),stopwords=FALSE))
as.matrix(contMat)
}
